Change default behavior of "list" command
Now, the "list" command just displays each identifier once instead of
grouping them by their template file. The old behavior is available
through the "--group" option. This change was made because the primary
purpose of the command is to quickly show users which identifiers are
missing from the config files.