Change behavior of "sync" command
Previously, source files that had been modified since the last sync
would be ignored. The problem with this behavior is that syncing a
second time would cause them to be overwritten anyways. Now, the time of
the last sync is tracked for each source file independently, so a source
file that is modified by the user will be ignored indefinitely until it
is synced with the '--overwrite' option.